pointers: detect a change of pointer-device used and set "touchscreen-pointer" as found in xfconf
To detect pointer device in use the DevicePresence-, DeviceMotionNotify- and
DeviceButtonPress-events are handled. Tests showed that by repopenning and
closing the pointer devices (e.g when another setting was modified), the events
stopped working. Therefore all pointer devices are opened only once and managed
in a hash table.

Upstream-Status: Submitted [1]

[1] https://bugzilla.xfce.org/show_bug.cgi?id=9474

Signed-off-by: Andreas Müller <schnitzeltony@googlemail.com>